chore: move EndpointV2 interface to types
