+++ packages/util-endpoints/src/utils/evaluateTreeRule.ts
@@ -1,8 +1,10 @@
-import { Endpoint, EvaluateOptions, TreeRuleObject } from "../types";
+import { EndpointV2 } from "@aws-sdk/types";
+
+import { EvaluateOptions, TreeRuleObject } from "../types";
 import { evaluateConditions } from "./evaluateConditions";
 import { evaluateRules } from "./evaluateRules";
 
-export const evaluateTreeRule = (treeRule: TreeRuleObject, options: EvaluateOptions): Endpoint | undefined => {
+export const evaluateTreeRule = (treeRule: TreeRuleObject, options: EvaluateOptions): EndpointV2 | undefined => {
   const { conditions, rules } = treeRule;
 
   const { result, referenceRecord } = evaluateConditions(conditions, options);
